Implemented kernel components

+++ CUDA/gramschmidt.h
@@ -34,4 +34,4 @@
 #  endif
 # endif /* !N */
 
-using DATA_TYPE = double;
+using DATA_TYPE = float;
